requirements.txt
aerich==0.7.1
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.2
click==8.1.3
dictdiffer==0.9.0
fastapi==0.94.1
h11==0.14.0
idna==3.4
iso8601==1.1.0
pydantic==1.10.6
PyMySQL==1.0.2
pypika-tortoise==0.1.6
pytz==2022.7.1
sniffio==1.3.0
starlette==0.26.1
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.0

redis~=4.5.4